AB-582: Update Media Embed for Display Inline.
Created the view mode for the display inline media types needed.
Added the theme suggestion for the figure caption wrapper to be able to select and give a class to it.
Updated Site Studio custom styles to properly have the content in the embed be inline.
Updated the WYSIWYG editor formats to allow the new view mode to be selectable.